requirements.txt
stravalib
python-dotenv
requests 